production/rest_api/requirements.txt
asgiref==3.2.10
django==3.1.1
pytz==2020.1
sqlparse==0.3.1
djangorestframework==3.11.1
django-rest-swagger==2.2.0
#mysqlclient==2.0.1
drf-yasg==1.17.1
humanfriendly==8.2
django-cors-headers==3.5.0
wheel==0.35.1
shortuuid==1.0.1
xlrd==1.2.0
gunicorn==20.0.4
django-rest-swagger==2.2.0
psycopg2==2.8.6
